Updated main.py - Working change in distribution over intervals
This main.py now has temporal structure, distribution analysis and change in distribution over intervals for each of the four selected stocks from the four groups after clustering. The four selected stocks are not random, nor are they explicitly chosen by me, they are just what is used temporarily.

The main goal now is to make sure that the stocks chosen from each group (1 for each of the 4 groups) is either completely random or purposefully chosen by mean with a reason.

This doesn't have any machine learning code in it as of yet. That will be the next step, build the model, train the model and use the model for forecast prediction of stock prices in 1 week, 2 weeks and 1 month.